Use design like site upload deletions for user pic
This commit makes use of an rails feature not yet added to
it: Allowing data-method to be acompanied by data-params
even if data-remote is not true. This way we can send a
synchronous request with request data.

To get it to work check out https://github.com/sternenseemann/rails/tree/method-params
and add change the "@rails/ujs" line in package.json to something
like "@rails/ujs": "link:/home/lukas/src/ruby/rails/actionview"
and run yarn install --force.

Rails Discuss entry for the feature is here:
https://discuss.rubyonrails.org/t/allow-data-params-for-ujs-links-that-dont-have-data-remote-true/73992

rspec tests succeed for this commit.