Match avatar/header clearing with site upload UI
For this we replace the buttons with <a>s which send the deletion
request using UJS. Unfortunately @rails/ujs only supports a
request body when using data-remote=true which makes the request
asynchronous. In this case we don't want this, since the changes
made by the request will only be visible after reloading the
settings page. To address this this commit adds
mastodon/utils/sync-ujs.js which implements exactly this feature.